+++ src/trx/game/inventory_ring/control.h
@@ -9,6 +9,7 @@ void InvRing_Close(INV_RING *ring);
 
 GF_COMMAND InvRing_Control(INV_RING *ring);
 bool InvRing_IsRingAvailable(RING_TYPE ring_type);
+INV_RING *InvRing_GetActiveRing(void);
 
 void InvRing_AdjustMusicVolume(const INV_RING *ring);
 void InvRing_SetRequestedObjectID(OBJECT_ID obj_id);